demo_rest_service(flask)/requirements.txt
click
Flask
gunicorn
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
Werkzeug
zipp
flask-cors
Flask-SQLAlchemy
pymysql
python-dotenv
cryptography
sqlalchemy